GH-323: [refactor] Move event related classes under the same package
