+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/controller/event/EventController.java
@@ -1,8 +1,7 @@
 package app.sportahub.eventservice.controller.event;
 
 import app.sportahub.eventservice.dto.request.event.EventRequest;
-import app.sportahub.eventservice.dto.request.EventCancellationRequest;
-import app.sportahub.eventservice.dto.request.EventRequest;
+import app.sportahub.eventservice.dto.request.event.EventCancellationRequest;
 import app.sportahub.eventservice.dto.response.EventResponse;
 import app.sportahub.eventservice.dto.response.ParticipantResponse;
 import app.sportahub.eventservice.enums.EventSortingField;
